fix(application): split seo.mts and restore generated env icon
Split the monolithic seo.mts into three focused files:
- seo-html.mts: all HTML injection helpers (meta tags, OG, JSON-LD,
  canonical, escape utilities)
- seo-sitemap.mts: sitemap XML builders (urlset, url elements,
  sitemap-index)
- seo.mts: plugin definition only, now also extracts the additional
  sitemap writing logic into a named writeAdditionalSitemaps function
  and the entry collection into buildSitemapEntries

Removes the ASCII-banner comments that were separating sections since
file boundaries now convey that structure.

Also restores the SVG data-URI icon in environment.g.d.ts that the
build regenerates with a placeholder when the icon host is not
accessible.
